Better support for ignoring pushes by instructors.
Pushes are now also ignored *after* a solution has been registered, using a list of ignored pushers. Since the general Git API cannot be used to determine who pushed something, more works needs to be done in a GitLab-specific way. But this also simplifies the rest of the code, e.g., no need for LatestCommitFetcher anymore.